requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
flask-jwt-extended==4.6.0
flask-cors==4.0.0
pymongo==4.6.1
werkzeug==3.0.1
pillow==10.2.0
pydicom==2.4.4
requests==2.31.0
python-dotenv==1.0.0
dnspython==2.8.0
